backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
